refactor: feedback changes to several files
